improve: add `makemirror` option
Sometimes we'd like to make a local mirror and that mirror must contain
packages we need only (but not an official mirror content).
It is supposed we've already cached that packages and we can run
`genini` (https://sourceware.org/viewvc/cygwin-apps/genini/)
script to generate new setup.ini file. So, we need `setup.hint` files
for an every package.
Let's automate getting that files for us from a mirror.